@@ -28,30 +28,27 @@
   },
   "homepage": "https://github.com/mohuk/ng-http#readme",
   "devDependencies": {
-    "@angular/compiler-cli": "^2.4.0",
-    "@angular/platform-browser": "^2.4.0",
-    "@angular/platform-browser-dynamic": "^2.4.0",
-    "@angular/compiler": "^2.4.0",
-    "@types/core-js": "^0.9.34",
-    "@types/jasmine": "^2.2.34",
-    "@types/node": "^6.0.41",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/common": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "typescript": "2.7.2",
+    "rxjs": "6.0.0",
+    "zone.js": "0.8.26",
+    "core-js": "2.5.4",
+    "@types/node": "8.9.4",
+    "@types/core-js": "2.5.0",
     "angular2-template-loader": "^0.5.0",
     "html-loader": "^0.4.4",
     "html-webpack-plugin": "^2.22.0",
-    "rimraf": "^2.5.4",
+    "rimraf": "^2.6.2",
     "ts-loader": "^0.8.2",
-    "typescript": "^2.0.10",
     "webpack": "^1.13.2",
     "webpack-dev-server": "^1.16.1",
-    "webpack-merge": "^0.14.1"
-  },
-  "peerDependencies": {
-    "@angular/common": "^2.4.0",
-    "@angular/core": "^2.4.0",
-    "@angular/http": "^2.4.0",
-    "rxjs": "^5.0.0-beta.12",
-    "zone.js": "^0.6.26",
-    "core-js": "^2.4.1",
+    "webpack-merge": "^0.14.1",
     "reflect-metadata": "^0.1.9"
   }
 }
